move tests to feature instead
